arbitrage-bot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
ccxt==4.1.22
pandas==2.1.4
numpy==1.26.2
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.2
